OK-831: Järjestelty sijoittelun-tulokset-state apufunktioita ja tyyppejä uudestaan ja parannettu nimiä
