+++ src/app/haku/[oid]/hakukohde/[hakukohde]/sijoittelun-tulokset/components/ilmoittautumis-cell.tsx
@@ -6,7 +6,7 @@ import {
 } from '@/app/lib/types/sijoittelu-types';
 import { SelectChangeEvent } from '@mui/material';
 import { isIlmoittautuminenPossible } from '@/app/lib/sijoittelun-tulokset-utils';
-import { SijoittelunTulosChangeParams } from '../lib/sijoittelun-tulokset-state-types';
+import { SijoittelunTulosChangeParams } from '../lib/sijoittelun-tulokset-state';
 import { useIlmoittautumisTilaOptions } from '@/app/hooks/useIlmoittautumisTilaOptions';
 
 export const IlmoittautumisCell = ({
