+++ src/app/haku/[oid]/hakukohde/[hakukohde]/sijoittelun-tulokset/components/maksu-cell.tsx
@@ -3,7 +3,7 @@ import { useTranslations } from '@/app/hooks/useTranslations';
 import { MaksunTila } from '@/app/lib/types/ataru-types';
 import { SijoittelunHakemusValintatiedoilla } from '@/app/lib/types/sijoittelu-types';
 import { SelectChangeEvent } from '@mui/material';
-import { SijoittelunTulosChangeParams } from '../lib/sijoittelun-tulokset-state-types';
+import { SijoittelunTulosChangeParams } from '../lib/sijoittelun-tulokset-state';
 
 export const MaksuCell = ({
   hakemus,
