+++ src/app/haku/[oid]/hakukohde/[hakukohde]/sijoittelun-tulokset/components/sijoittelun-tulos-form.tsx
@@ -5,14 +5,12 @@ import useToaster from '@/app/hooks/useToaster';
 import { useActorRef } from '@xstate/react';
 import { Box } from '@mui/material';
 import { SijoittelunTuloksetActions } from './sijoittelun-tulos-actions';
-import {
-  createSijoittelunTuloksetMachine,
-  useIsDirtySijoittelunTulos,
-} from '../lib/sijoittelun-tulokset-state';
+import { createSijoittelunTuloksetMachine } from '../lib/sijoittelun-tulokset-state';
 import { SijoitteluajonValintatapajonoValintatiedoilla } from '@/app/lib/types/sijoittelu-types';
 import { Haku, Hakukohde } from '@/app/lib/types/kouta-types';
 import { SijoittelunTulosTable } from './sijoittelun-tulos-table';
 import { useConfirmChangesBeforeNavigation } from '@/app/hooks/useConfirmChangesBeforeNavigation';
+import { useIsDirtySijoittelunTulos } from '../lib/sijoittelun-tulokset-state-utils';
 
 type SijoittelunTuloksetFormParams = {
   valintatapajono: SijoitteluajonValintatapajonoValintatiedoilla;
